pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "simpla-loop"
version = "0.1.0"
description = "Educational framework for agentic loop patterns"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "Michał Pawłowski", email = "misza222@gmail.com"}
]
requires-python = ">=3.12"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Education",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "instructor>=1.0",
    "openai>=1.0",
    "python-dotenv>=1.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov",
    "mypy>=1.0",
    "ruff>=0.1.0",
    "instructor>=1.0",
    "openai>=1.0",
    "python-dotenv>=1.0",
    "pydantic>=2.0",
]
docs = [
    "mkdocs",
    "mkdocstrings[python]",
]

[tool.hatch.build.targets.wheel]
packages = ["src/simpla_loop"]

[tool.ruff]
target-version = "py310"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=simpla_loop --cov-report=term-missing"
